// front-end/src/types/TypeAssessment.tsx
export type UrgencyStatus = "NORMAL" | "URGENT" | "EMERGENCY";
// ...
export interface UserBasicInfo {
  id: string;
  firstname: string;
  lastname: string;
  email: string;
  employeeId?: string;
}

export interface Mechanic {
  id: string;
  employeeId: string;
  userName: string;
  firstname: string;
  lastname: string;
  email: string;
  role: string;
  imageUrl: string | null;
  sectionId?: string;
  section?: Section;
  activeJobsCount?: number;
}

export interface SparePart {
  id: number;
  code: string;
  name: string;
  qtyInStock: number;
  price: number;
  unit: string;
  minStock?: number;
  groupId?: number;
}


// REPAIR MODELS & RESPONSES
export interface RepairListItem {
  id: string;
  jobNo: string;
  symptom: string;
  urgencyStatus: UrgencyStatus;
  createdAt: string;
  isReassessment?: boolean;
  rejectionReason?: string | null;
  mechanicRepairs?: MechanicRepair[];
  asset?: {
    id: string;
    name: string;
    noid: string;
    type?: {
      id: number;
      name: string;
    };
  };
}

export interface MechanicRepair {
  id: number;
  jobId?: string;
  userId: string;
  createdAt?: string;
  updatedAt?: string;
  deletedAt?: string | null;
  user?: Mechanic | UserBasicInfo;
}